back/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
Flask-Cors
python-dotenv
google-generativeai